start backslashes + print debug

+++ srcs/parse.c
@@ -6,7 +6,7 @@
 /*   By: hthomas <hthomas@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/09/17 15:52:09 by hthomas           #+#    #+#             */
-/*   Updated: 2020/10/19 11:52:01 by hthomas          ###   ########.fr       */
+/*   Updated: 2020/10/19 16:33:44 by hthomas          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -39,26 +39,32 @@ void	err_code(t_list_cmd *cmd, char **envp)
 	return ;
 }
 
-// int 	deal_backslash(t_list_command *cmd, char **envp)
-// {
-// 	int		i;
+int 	delete_backslashes(t_list_cmd *cmd, char **envp)
+{
+	int		i;
 
-// 	while (cmd)
-// 	{
-// 		i = 0;
-// 		while (cmd->str && cmd->str[i])
-// 		{
-// 			if (cmd->str[i] == '\\' && ...)
-// 			{
-// 				//? i++;
-// 				//! to do
-// 			}
-// 			i++;
-// 		}
-// 		cmd = cmd->next;
-// 	}
-// 	return (0);
-// }
+	while (cmd)
+	{
+		i = 0;
+		while (cmd->str && !(cmd->flags & F_SIMPLE_QUOTES) && cmd->str[i])
+		{
+			if (cmd->str[i] == '\\')
+			{
+				ft_printf("backslash: %c\n", cmd->str[i]);
+				if (cmd->str[i + 1])
+					ft_strcpy(&cmd->str[i], &cmd->str[i + 1]);
+				else
+				{
+					ft_putstr(cmd->str);
+					return (FAILURE);
+				}
+			}
+			i++;
+		}
+		cmd = cmd->next;
+	}
+	return (SUCCESS);
+}
 
 int		replace_dollar_and_tild(t_list_cmd *cmd, char **envp)
 {
@@ -70,14 +76,14 @@ int		replace_dollar_and_tild(t_list_cmd *cmd, char **envp)
 		i = 0;
 		while (cmd->str && cmd->str[i])
 		{
-			if (cmd->str[i] == '$' && !(cmd->flags & F_SIMPLE_QUOTES) && cmd->str[i + 1] > 32)
+			if (cmd->str[i] == '$' && !escaped(cmd->str, i) && !(cmd->flags & F_SIMPLE_QUOTES) && cmd->str[i + 1] > 32)
 			{
 				if (cmd->str[i + 1] == '?')
 					err_code(cmd, envp);
 				else
 					replace_all_var_env(cmd, envp, i);
 			}
-			else if (cmd->str[i] == '~' && !in_quotes(cmd) && (!cmd->str[i + 1] || cmd->str[i + 1] == '/'))
+			else if (cmd->str[i] == '~' && !escaped(cmd->str, i) && !in_quotes(cmd) && (!cmd->str[i + 1] || cmd->str[i + 1] == '/'))
 			{
 				tmp = cmd->str;
 				cmd->str = ft_strdup(&find_var_env(envp, "HOME=")[5]);
@@ -87,7 +93,7 @@ int		replace_dollar_and_tild(t_list_cmd *cmd, char **envp)
 		}
 		cmd = cmd->next;
 	}
-	return (0);
+	return (SUCCESS);
 }
 
 void	delete_empty_elements(t_list_cmd *cmd)
@@ -112,14 +118,14 @@ int		input_to_command(char *input, t_list_cmd **cmd)
 	init_par(&par);
 	while (input[par.i])
 	{
-		if (input[par.i] == '\'' && !par.in_double && !escaped(input, par.i))
+		if (input[par.i] == '\'' && !par.in_double)
 			simple_quotes(input, cmd, &par);
 		else if (input[par.i] == '\"' && !par.in_simple && !escaped(input, par.i))
 			double_quotes(input, cmd, &par);
 		else if (is_separator(&input[par.i]) && !escaped(input, par.i) && !par.in_simple && !par.in_double)
 			separator(input, cmd, &par);
 		// si je suis sur un mot et hors de quotes
-		else if (!ft_in_charset(input[par.i], WHITESPACES) && !par.in_simple && !par.in_double)
+		else if ((!ft_in_charset(input[par.i], WHITESPACES) || escaped(input, par.i)) && !par.in_simple && !par.in_double)
 			end_word(input, cmd, &par);
 		par.i++;
 	}
@@ -168,8 +174,21 @@ int		parse_input(char *input, t_list_line **lst_line, char **envp)
 	cmd = NULL;
 	if (input_to_command(input, &cmd))
 		return (FAILURE);
-	// deal_backslash((**lst_line)->cmd, envp);
+
+	ft_putstr("-----------------CMD:\n");
+	t_list_cmd	*copy = cmd;
+		while (copy)
+		{
+			ft_putstr(copy->str);
+			ft_putstr("\n");
+			copy = copy->next;
+		}
+
+
 	replace_dollar_and_tild(cmd, envp);
+	if (delete_backslashes(cmd, envp))
+		return (FAILURE);
+	ft_putstr("Ready for execution.................\n");
 	delete_empty_elements(cmd);
 	l_lst_add_back(lst_line, l_lst_new(cmd, '\0'));
 	if (split_cmd(lst_line, cmd, 0))
@@ -190,7 +209,5 @@ int		parse_input(char *input, t_list_line **lst_line, char **envp)
 	// 	copy = (copy)->next;
 	// }
 	// ft_putstr("!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!\n");
-
-	//c_lst_clear(cmd);
 	return (SUCCESS);
 }
